+++ src/test/java/com/salesforce/dynamodbv2/UpdateTest.java
@@ -121,6 +121,37 @@ class UpdateTest {
         });
     }
 
+
+    @ParameterizedTest(name = "{arguments}")
+    @ArgumentsSource(DefaultArgumentProvider.class)
+    void updateConditionalOnGsiHkSuccess(TestArgument testArgument) {
+        testArgument.forEachOrgContext(org -> {
+            testArgument.getAmazonDynamoDb().updateItem(new UpdateItemRequest()
+                .withTableName(TABLE3)
+                .withKey(ItemBuilder.builder(testArgument.getHashKeyAttrType(), HASH_KEY_VALUE)
+                    .rangeKey(S, RANGE_KEY_S_VALUE)
+                    .build())
+                .withUpdateExpression("set #indexField = :newValue")
+                /*
+                 * Using ImmutableMap's for expressionAttributeNames and expressionAttributeValues here to
+                 * intentionally test that they are properly handled.
+                 */
+                .withExpressionAttributeNames(ImmutableMap.of("#indexField", INDEX_FIELD))
+                .withExpressionAttributeValues(ImmutableMap.of(":newValue",
+                    createStringAttribute(INDEX_FIELD_VALUE + TABLE3 + org + "Updated"))));
+            assertThat(getItem(testArgument.getAmazonDynamoDb(),
+                TABLE3,
+                HASH_KEY_VALUE,
+                testArgument.getHashKeyAttrType(),
+                Optional.of(RANGE_KEY_S_VALUE)),
+                is(ItemBuilder.builder(testArgument.getHashKeyAttrType(), HASH_KEY_VALUE)
+                    .someField(S, SOME_FIELD_VALUE + TABLE3 + org)
+                    .indexField(S, INDEX_FIELD_VALUE + TABLE3 + org + "Updated")
+                    .rangeKey(S, RANGE_KEY_S_VALUE)
+                    .build()));
+        });
+    }
+
     @ParameterizedTest(name = "{arguments}")
     @ArgumentsSource(DefaultArgumentProvider.class)
     void updateConditionalOnHkWithLiteralsSuccess(TestArgument testArgument) {